Get the project on its feet.
This commit is a little big to talk about every change so I'll give an overview here.
Basically, the project is now in a semi-working state. It can play any (correctly formatted) soundscape in the browser that doesn't build off of another soundscape. It can play looping sounds and randomized sounds. There's a lot going on here so I'll do the rest in bullet point form.
Right now there are all the soundscapes from hl2, and all the hl2 ambient sounds. This covers most of the sounds needed in the soundscape scripts, but a couple of them use sounds from the npc or some other misc folder (like button).
The site will play any looping sounds and randomized sounds at the volume specified in the script.
The sounds will be played at the same volume everywhere
There are a couple problems that I can see coming up, such as getting pitches to work since there is no way to alter the pitch of a sound using the built-in audio system.
Everything else should be self-explainitory.