Begin replacing web audio with Web Audio API
This is two steps backwards, one step forward. In order to accomplish some of the more advanced parts of a soundscape, such as pitch and maybe panning/positioning in the future, a different audio system has to be used. This commit is the first of many that will eventually replace the basic web audio with the more advanced Web Audio API. I'm doing this step by step so some features are currently lost at this point in time. They will be rebuilt in the new system.

A huge help in this is "screw" by `dumbmatter` found here: https://github.com/dumbmatter/screw/ which in turn uses soundtouch found here: https://github.com/also/soundtouch-js

It's a great base, and I plan on modifying it a good amount to make it work for this project.