Fix clicking and playrandom sound truncation problem
Adds a 0.03s fadeout when stopping a sound before actually stopping the sound. This should remove any clicks that come from bad timing. Also add a slight delay to emitting the end of a sound. There are still some issues with very short sounds like water drips and especially noticeable because of the high decay time