Update readme to explain certain soundscape rule implementations
Part duex!
This time for attenuation and soundlevel. Since space is simulated with the position/origin keyvalues, and the sound volume is lowered because of it, it's basically already attenuation built in at the most basic level.

Also sometimes Valve's soundlevels are wack. Sometimes water dripping is louder than a vacuum cleaner according to it's attenuation value.