Implement Game Selector; part 2
This commit finishes up what the previous one started by including the soundscapes for the other games and all the associated sound files. There are a lot of sound files. To make this not inflate the repo like a balloon, this commit also introduces LFS for all sound files. This is my first time implementing FLS partway into a repo, so hopefully this works out.

Theoretically this commit should work just fine, but the next step is to do a whole bunch of testing to make sure.