Begin styling the app
Now that most of the functionality is complete, it's time to move on to making it look presentable. The two remaining features I'm looking to add for sure will be dependent on how the final thing looks, so I think nows the time to make some progress.

This isn't final, of course. I'm still planning on completely reworking the center content portion to not only display the current soundscape in a better fashion, but I'm also looking into finding a way to bring up a relevant image as a background (i.e. show the citadel if a citadel soundscape is being played etc). I don't want to take and include hundreds of pictures though, so we'll see if I can find a nice way to do it.

Only thing that was removed in this commit is the ability to stop playback completely. There isn't a blank option in the dropdown menu because it was kinda ugly and I'm planning on making a working stop button whose functionality is more obvious than just selecting nothing from a dropdown. That's coming next.