Make Favorite and Stop buttons work
Now you can click the star next to a soundscript to favorite/unfavorite it. Favorited soundscapes will show up in the "FAVORITE" game, as well as appear with a star next to their name in the list of soundscapes. Favorited soundscapes are saved in a cookie so that they will be persistent across refreshes.
The stop button stops playback without changing the current soundscape.