v1.1.0
Completely rework file system, prune unused sounds, add volume cookie, dramatically reduce repo size

This commit reduces the size of the repo by about 90% by removing all sounds from LFS and the repo, pruning the unused sounds (about 1GB(!) of them), and moving them it's own archive (which will probably be kept in the releases area). This zip must be extracted locally for the site to work.
This commit also saves the selected volume as a cookie for persistence.